Port tests to DNX with xUnit
