+++ Dapper.FluentMap.sln
@@ -13,6 +13,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Dapper.FluentMap", "src\Dapper.FluentMap\Dapper.FluentMap.xproj", "{3ECF719B-E39F-40E5-8B4E-DC1F8E429E32}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Dapper.FluentMap.Tests", "test\Dapper.FluentMap.Tests\Dapper.FluentMap.Tests.xproj", "{FF3FAE2A-EB78-46CC-B049-ED4F5123BC4B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,11 +25,16 @@ Global
 		{3ECF719B-E39F-40E5-8B4E-DC1F8E429E32}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3ECF719B-E39F-40E5-8B4E-DC1F8E429E32}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3ECF719B-E39F-40E5-8B4E-DC1F8E429E32}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FF3FAE2A-EB78-46CC-B049-ED4F5123BC4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF3FAE2A-EB78-46CC-B049-ED4F5123BC4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FF3FAE2A-EB78-46CC-B049-ED4F5123BC4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF3FAE2A-EB78-46CC-B049-ED4F5123BC4B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{3ECF719B-E39F-40E5-8B4E-DC1F8E429E32} = {580E3446-6579-4414-9875-970849E635E5}
+		{FF3FAE2A-EB78-46CC-B049-ED4F5123BC4B} = {580E3446-6579-4414-9875-970849E635E5}
 	EndGlobalSection
 EndGlobal
