+++ test/Dapper.FluentMap.Tests/TestEntity.cs
