+++ test/Dapper.FluentMap.Tests/project.json
@@ -1,0 +1,15 @@
+﻿{
+  "dependencies": {
+    "Dapper.FluentMap": "2.0.0-beta-*",
+    "xunit": "2.1.0",
+    "xunit.runner.dnx": "2.1.0-rc1-build204"
+  },
+
+  "frameworks": {
+    "dnx451": { }
+  },
+
+  "commands": {
+    "test": "xunit.runner.dnx"
+  }
+}
